+++ clang/lib/Sema/SemaExpr.cpp
@@ -14578,21 +14578,51 @@ ExprResult Sema::CreateBuiltinBinOp(SourceLocation OpLoc,
                                     BinaryOperatorKind Opc,
                                     Expr *LHSExpr, Expr *RHSExpr) {
   if (getLangOpts().CPlusPlus11 && isa<InitListExpr>(RHSExpr)) {
-    // The syntax only allows initializer lists on the RHS of assignment,
-    // so we don't need to worry about accepting invalid code for
-    // non-assignment operators.
-    // C++11 5.17p9:
-    //   The meaning of x = {v} [...] is that of x = T(v) [...]. The meaning
-    //   of x = {} is x = T().
-    InitializationKind Kind = InitializationKind::CreateDirectList(
-        RHSExpr->getBeginLoc(), RHSExpr->getBeginLoc(), RHSExpr->getEndLoc());
+    // C++11 [expr.ass]p9, per CWG2768:
+    //   A braced-init-list B may appear on the right-hand side of
+    //    - an assignment to a scalar of type T, in which case B shall have at
+    //      most a single element. The meaning of x = B is x = t, where t is an
+    //      invented temporary variable declared and initialized as T t = B.
+    if (Opc != BO_Assign) {
+      // A compound assignment like `i += {0}` is equivalent to `i = i + {0}`,
+      // which is a parsing error
+      assert(BinaryOperator::isCompoundAssignmentOp(Opc) &&
+             "Non-assignment binary operator with braced-init-list should not "
+             "be parsed");
+      Diag(OpLoc, diag::err_init_list_bin_op)
+          << 1
+          << BinaryOperator::getOpcodeStr(
+                 BinaryOperator::getOpForCompoundAssignment(Opc))
+          << getExprRange(RHSExpr);
+      return ExprError();
+    }
+
+    QualType LHSTy = LHSExpr->getType();
+    assert(!LHSTy->isDependentType() &&
+           "Should not have tried to create a builtin binary operator");
+    // Extend this to be done for non-scalars too. This is so extension types
+    // like vectors can be assigned to with an initializer list
+    if (LHSTy->isArrayType()) {
+      // Except arrays, which we know this will fail for, but fail early to
+      // prevent trying to convert the initializer list elements to the array
+      // type
+      Diag(OpLoc, diag::err_typecheck_array_not_modifiable_lvalue)
+          << LHSTy << getExprRange(LHSExpr);
+      return ExprError();
+    }
+    InitializationKind Kind =
+        InitializationKind::CreateCopy(RHSExpr->getBeginLoc(), OpLoc);
     InitializedEntity Entity =
         InitializedEntity::InitializeTemporary(LHSExpr->getType());
     InitializationSequence InitSeq(*this, Entity, Kind, RHSExpr);
-    ExprResult Init = InitSeq.Perform(*this, Entity, Kind, RHSExpr);
-    if (Init.isInvalid())
-      return Init;
-    RHSExpr = Init.get();
+    ExprResult InventedTemporary =
+        InitSeq.Perform(*this, Entity, Kind, RHSExpr);
+    if (InventedTemporary.isInvalid())
+      return InventedTemporary;
+    // The "at most a single element" condition should be checked by this
+    // initialization succeeding, but allow multiple initializers for the
+    // extension type _Complex T / [[gnu::vector_size]]
+    RHSExpr = InventedTemporary.get();
   }
 
   ExprResult LHS = LHSExpr, RHS = RHSExpr;